# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Đây là tài liệu Swagger cho hệ thống.",
        "title": "API DATN 2025",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Kin",
            "email": "trannguyentrungkien1006@gmail.com"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/posts": {
            "get": {
                "description": "API bao gồm cả lọc, phân trang và sắp xếp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Get posts",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "description": "Current page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 10,
                        "description": "Number record of page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "authorName, title, createdAt",
                        "description": "Sort column",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "ASC, DESC",
                        "description": "Sort type",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "1, 2, 3",
                        "description": "Pending:1, Rejected:2, Approved:3",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "1, 2, 3, 4",
                        "description": "GiveAwayOldItem:1, FoundItem:2, SeekLoseItem:3, Other:4",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Trường lọc (vd: email, fullName)",
                        "name": "searchBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Giá trị lọc (vd:abc@gmail.com, John Doe)",
                        "name": "searchValue",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/postdto.GetAdminPostResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            },
            "post": {
                "description": "API tạo mới một post và trả về thông tin post + user + JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Create a new post",
                "parameters": [
                    {
                        "description": "Post creation payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postdto.CreatePostRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/postdto.CreatePostResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            }
        },
        "/posts/{postID}": {
            "get": {
                "description": "API lấy bài viết theo id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Get detail post",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID post",
                        "name": "postID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/postdto.GetDetailPostResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            },
            "patch": {
                "description": "API cập nhật bài viết kết hợp với patch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Update posts",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID post",
                        "name": "postID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update post info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postdto.UpdatePostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated post successfully",
                        "schema": {
                            "$ref": "#/definitions/postdto.UpdatePostResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "enums.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "INVALID_QUERY_PARAM"
                },
                "message": {
                    "type": "string",
                    "example": "Name is require"
                }
            }
        },
        "enums.PostStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-comments": {
                "PostStatusAll": "0 Người dùng đã gửi yêu cầu duyệt bài và đang chờ xử lý",
                "PostStatusApproved": "3 Admin duyệt",
                "PostStatusPending": "1 Người dùng đã gửi yêu cầu duyệt bài và đang chờ xử lý",
                "PostStatusRejected": "2 Admin đã từ chối"
            },
            "x-enum-varnames": [
                "PostStatusAll",
                "PostStatusPending",
                "PostStatusRejected",
                "PostStatusApproved"
            ]
        },
        "enums.PostType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-comments": {
                "PostTypeAll": "0 all",
                "PostTypeFoundItem": "2 nhặt được đồ",
                "PostTypeGiveAwayOldItem": "1 tặng đồ cũ",
                "PostTypeOther": "4 khác",
                "PostTypeSeekLoseItem": "3 tìm kiếm đồ"
            },
            "x-enum-varnames": [
                "PostTypeAll",
                "PostTypeGiveAwayOldItem",
                "PostTypeFoundItem",
                "PostTypeSeekLoseItem",
                "PostTypeOther"
            ]
        },
...
        "postdto.AdminPostDTO": {
            "type": "object",
            "properties": {
                "authorName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "postdto.CreatePostRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "strbase64",
                        " strbase64"
                    ]
                },
                "info": {
                    "type": "string"
                },
                "newItems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/postdto.NewItemsPost"
                    }
                },
                "oldItems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/postdto.OldItemsPost"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "Tôi muốn tìm đồ thất lạc"
                },
                "type": {
                    "description": "Email       string         `json:\"email\" example:\"john@gmail.com\"`\nFullName    string         `json:\"fullName\" example:\"John Doe\"`\nPhoneNumber string         `json:\"phoneNumber\" example:\"0123456789\"` // true: anonymous, false: not anonymous",
                    "enum": [
                        1,
                        2,
                        3,
                        4
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.PostType"
                        }
                    ],
                    "example": 1
                }
            }
        },
        "postdto.CreatePostResponseWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "postdto.DetailPostDTO": {
            "type": "object",
            "properties": {
                "authorID": {
                    "type": "integer"
                },
                "authorName": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "info": {
                    "type": "string"
                },
                "interests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/postdto.InterestDTO"
                    }
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/postdto.DetailPostItemDTO"
                    }
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "postdto.DetailPostItemDTO": {
            "type": "object",
            "properties": {
                "categoryID": {
                    "type": "integer"
                },
                "categoryName": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "itemID": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "postdto.GetAdminPostResponse": {
            "type": "object",
            "properties": {
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/postdto.AdminPostDTO"
                    }
                },
                "totalPage": {
                    "type": "integer"
                }
            }
        },
        "postdto.GetAdminPostResponseWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/postdto.GetAdminPostResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "postdto.GetDetailPostResponse": {
            "type": "object",
            "properties": {
                "post": {
                    "$ref": "#/definitions/postdto.DetailPostDTO"
                }
            }
        },
        "postdto.GetDetailPostResponseWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/postdto.GetDetailPostResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "postdto.InterestDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "postID": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "userAvatar": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
        "postdto.NewItemsPost": {
            "type": "object",
            "required": [
                "categoryID",
                "name"
            ],
            "properties": {
                "categoryID": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "postdto.OldItemsPost": {
            "type": "object",
            "required": [
                "itemID"
            ],
            "properties": {
                "itemID": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "postdto.UpdatePostRequest": {
            "type": "object",
            "properties": {
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "strbase64",
                        " strbase64"
                    ]
                },
                "info": {
                    "type": "string"
                },
                "status": {
                    "enum": [
                        1,
                        2,
                        3
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.PostStatus"
                        }
                    ]
                },
                "title": {
                    "type": "string",
                    "example": "Tôi muốn tìm đồ thất lạc"
                }
            }
        },
        "postdto.UpdatePostResponseWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}